Added new quadcopter pilot node to introuce latency
Now the teleop from gazebo directly is removed. We only use script which
uses pyinput to send commands to the remote server via the
/model_movement_commands topic, which sends it over the network to the
UE, which in turn forwards it to the gazebo bridged /x3/cmd_vel topic.
This basically means that we are actually controlling the copter through
the network.