Added reinforcement learning path planner script dertails to readme
Just starting work for a reinforcement learning model that will populate
waypoints on the gazebo world, for the quadcopter pilot to follow. This
path will be adaptive, and keep changing depending on the path the
actual pilot takes, and will react to network changes (like unexpected
congestion). The objective is balanced in a way to reduce handovers,
have high RSRP throughout and taking the physically shortest path.