Added new callback for handovers, and new Gazebo plugin to monitor current cell
Now the ns3 script will automatically execture handovers based on some
preset hysteresis values, using the A3 algorithm. The UE also has a
handover complete listener, which notifies on the `/current_cell_id`
topic whenever there is a change. This topic is bridged with a topic
with the same name in the gazebo transport layer, and that topic is now
subscribed to inside the signal strength plugin, which now tracks the
currently connected towers/enodebs/cells.