Refactored ros node into separate class. Added periodic animation flush. Removed ros desktop utils
Now the netanim animation interface periodically flushes animations,
thanks to a scheduled call. I've also removed ros_desktop cuz the UI
seems more or less useless.  The big refactor is just moving the ROS
node into a separate cc file, and including it via a header. The CMAKE
changes are also just to include this new includes directory, and a new
cc file to be included in the executable.